requirements.txt
Django==1.10.1
django-filter==0.15.2
djangorestframework==3.4.7
gunicorn==19.6.0
Markdown==2.6.7
mysqlclient==1.3.7
